Move QuickBooks integration to Slim framework

+++ lib/Scat/Model/Config.php
@@ -35,7 +35,9 @@ class Config extends \Model {
 
   static public function forgetValue($name) {
     $row= \Model::factory('Config')->where('name', $name)->find_one();
-    $row->delete();
+    if ($row) {
+      $row->delete();
+    }
     unset($cache[$name]);
